requirements.txt
Flask_Babel==2.0.0
Flask_Bootstrap==3.3.7.1
alembic==1.7.4
SQLAlchemy==1.4.25
requests==2.26.0
XlsxWriter==1.1.5
WTForms==2.3.3
chardet==3.0.4
Flask_Migrate==3.1.0
click==8.0.3
Flask==2.0.2
numpy
Flask_Login==0.5.0
pyodbc==4.0.32
Flask_SQLAlchemy==2.5.1
pandas==0.24.1
Flask_WTF==0.15.1
Werkzeug==2.0.2
Flask_Moment==1.0.2
Flask_Mail==0.9.1
Flask_Admin==1.5.8
easygui==0.98.2
flask_table==0.5.0
guess_language==0.2
plac==1.3.3
PyJWT==2.2.0
python-dotenv==0.19.1
spacy==3.1.3
textract==1.6.4
